+++ src/main/java/com/mageddo/dnsproxyserver/config/Config.java
@@ -68,7 +68,7 @@ public class Config {
   private Path configPath;
 
   @NonNull
-  private Path resolvConfPath;
+  private String resolvConfPaths;
 
   public static SimpleServer.Protocol findDnsServerProtocol() {
     return SimpleServer.Protocol.BOTH;