+++ src/test/java/com/mageddo/dnsproxyserver/templates/ConfigTemplates.java
@@ -1,6 +1,7 @@
 package com.mageddo.dnsproxyserver.templates;
 
 import com.mageddo.dnsproxyserver.config.Config;
+import com.mageddo.dnsproxyserver.config.entrypoint.ConfigEnv;
 import com.mageddo.dnsproxyserver.config.entrypoint.LogLevel;
 
 import java.nio.file.Paths;
@@ -26,7 +27,7 @@ public class ConfigTemplates {
       .dnsServerPort(53)
       .domain("com")
       .logLevel(LogLevel.WARNING)
-      .resolvConfPath(Paths.get("/etc/resolv.conf"));
+      .resolvConfPaths(ConfigEnv.DEFAULT_RESOLV_CONF_PATH);
   }